pyproject.toml
[project]
name = "research-intelligence-platform"
version = "0.1.0"
description = "Multi-agent AI system for monitoring research literature"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}
authors = [
    {name = "Research Intelligence Team"}
]
keywords = ["ai", "agents", "research", "google-adk", "cloud-run"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
]

dependencies = [
    # Core ADK and Google AI
    "google-adk>=0.1.0",
    "google-genai>=1.0.0",  # New unified SDK (replaces google-generativeai)

    # Google Cloud Services
    "google-cloud-firestore>=2.14.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-pubsub>=2.19.0",
    "google-cloud-aiplatform>=1.38.0",

    # PDF Processing
    "PyMuPDF>=1.23.8",
    "pdfplumber>=0.10.3",

    # Basic ML/Data
    "numpy>=1.24.3,<2.0",

    # Web Framework
    "flask>=3.0.0",
    "flask-cors>=4.0.0",
    "gunicorn>=21.2.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "pydantic>=2.5.0",

    # arXiv API
    "arxiv>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "black>=23.12.0",
    "ruff>=0.1.8",
    "mypy>=1.7.0",
    "ipython>=8.18.0",
]

test = [
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
]

ml = [
    # ML/NLP for embeddings (Phase 2+)
    "sentence-transformers>=2.2.2",
    "torch>=2.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--cov=src",
    "--cov-report=term-missing",
    "--cov-report=html",
]

[tool.black]
line-length = 100
target-version = ["py39", "py310", "py311"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | venv
  | build
  | dist
)/
'''

[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
